chore: 고등학교 등록 및 버전, 연도 추가

+++ src/view/container/admin/SchoolContainer.tsx
@@ -23,8 +23,12 @@ const AdminSchoolContainer = ({
           const payload = await schoolRepository.findById(id);
           return payload;
         },
-        async addSchool() {},
+        async addSchool(req) {
+          const payload = await schoolRepository.save(req.data);
+          return payload;
+        },
         async deleteSchool(req) {
+          if (!req.id) return;
           await schoolRepository.delete(req.id);
         },
         async editSchool(req) {
